+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.1] - 2022-04-13
 ### Fixed
 - Next version not being detected correctly.
 - `NoSuchPathError` when creating a repo without a mapping specified.
@@ -61,7 +63,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Each version committed to local repo
 - Range operators (`..` and `...`)
 
-[Unreleased]: https://github.com/clabe45/shulkr/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/clabe45/shulkr/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/clabe45/shulkr/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/clabe45/shulkr/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/clabe45/shulkr/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/clabe45/shulkr/compare/v0.3.1...v0.3.2